Incrementatl checkin. Working on Magic.
Added "ParameterData" to Persistence, a variant data type. The
underlying data value is stored as a String- lists are supported,
saved as CSVs.

This is primarily to support Spell variants, which now take a list of
key/value ParameterData parameters, instead of a string list.

However, this will also directly support YML/flatfile data stores, as
well, and eventually I'd like to use this to replace the permissions
system if bukkit hasn't released theirs by the time I get far enough
to make that happen.