Add "concurrent" option to simulation batches
This will run the Life simulation on each cell at a time, instead of on
the entire space in one tick.

This tends to result in more explosive growth, and is used in the
hunters for greater mobility. I will have fun experimenting with it
when I get the time!