Had to test so much stuff to figure out this centering issue.
Turns out the way it works is actually really obvious.. the bounding box of the text just starts with the first char you write out, and ends at the end of the last char.

The implication though is that if you want things to center properly, if you have first write out whatever will be the left-most glyph, and make sure you end at the end of the right-most glyph.

For the hotbar + mana bar combo the end result is we have to figure out which one is longer and write that one out first, then write the other one surrounded by padding.

Have people really made a generalized rendering engine for this stuff?