Clean up Vivecraft integration
I am not 100% sure what the best way to handle targeting is, but in general this implicit pre-target is really a lot of trouble.

The way it is is configured is even worse, controlled by two separate parameters that are both overloads of completely unrelated and frequently-used parameters, like Volume actions' orient.

So hopefully there won't be any major negative consequences from bypassing it completely.