Fixed kill dealing less damage at higher levels
The kill spell deals less damage at higher levels. Specifically, its level one effect has "player_damage: 1000", but then the level two and level three effects have "player_damage:25" and "player_damage:30" in them, and both of those numbers are significantly less than 1000.

This should fix that. 

Also, the Kill spell only deals 30 (up to 40 with max level) damage against entities. I'm not sure if this is a bug or not, so I left it untouched, but it is something to look at.