Fix spells cast by mobs only damaging the targeted player.
Worldguard considers it PVP when a mob damages a player that was
targeting another player. This means that any (magic) mob casting
an AOE spell, or a spell that otherwise hits a player that it was
not targeting triggers this event, and does not do damage by
default.

Here at Magic inc. we don't that this behaviour makes sense for
spells, as mobs casting spells are typically highly customised,
and can do their own PVP/player-targeting checks. That is why we
will allow PVP when we are the ones that caused the damage.