Force-create custom data components when trying to write to a tag.
This is a compromise from the last commit that always force-created when trying to get a tag.

This is only necessary in 1.20.5 and up because the tag won't exist but shouldn't be created unless we intend to write to the item.

This fixes Selector menus since the slot tag wasn't getting written to menu items.