Feature - RectOfEffect action
Action which selects targets within a cuboid area based on a min and max relative position.
This action may still be in need of optimization as I have not tested for this specifically.
I am not fully confident I understood the 'addwork' part of the spell actions, this may require further inspection.